backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
Flask-Cors
psycopg2-binary
redis
python-dotenv
Flask-JWT-Extended
Werkzeug
